Rename private module kas_widgets::adapter::label to with_label
