+++ crates/kas-widgets/src/adapter/with_label.rs
